+++ pages/_app.tsx
@@ -16,9 +16,4 @@ const ObsidianPluginStatsApp = ({ Component, pageProps }) => {
   );
 }
 
-// export default withApplicationInsights({
-//   instrumentationKey: 'f34222fa-b4e2-4961-be30-1bfd720292bb',
-//   isEnabled: true //process.env.NODE_ENV === 'production'
-// })(ObsidianPluginStatsApp);
-
 export default ObsidianPluginStatsApp;